Class name changed
Updated the the class name according to the code style guide lines
